Date: Wed, 15 Sep 2004 23:26:42 -0500 From: Jonathan Atkins Subject: [SDL] [SDL_Mixer] external libmikmod patch
here's a great patch for the CVS of SDL_mixer.  It allows people to use the preinstalled version of libmikmod instead of the internal one.
I know that Gentoo users at least want this option, and I also see the logic in it.

It's all ./configure'able so the user can choose whether to use the included mikmod source, or just link with the external.
the new option is --enable-music-libmikmod
if it is disabled or not found, then it rolls off to default to the included mikmod source, like it did before.

This also implements the new SDL_LoadMUS_RW functionality with a libmikmod Player_LoadGeneric adaptor.
This element was tested with a revised playmus that uses RWops to load a module from a fake file with a module at a known offset (see the -o option).
I've attached that new file as well, but if you use it, you'll have to change Makefile.am yourself :)

The only thing that turns into a bogus function is the Mix_(Set/Get)SynchroValue functions, which now just set/get a static int, and print a warning to stderr.
I felt that functionality is not used very much, if at all...  and if used, should patched into libmikmod upstream instead, since it doesn't currently support it.

--HG--
extra : convert_revision : svn%3Ac70aab31-4412-0410-b14c-859654838e24/trunk%401443