Updated Xcode project from Eric Wing:
Meanwhile to bundle a framework transparently in the SDL_mixer framework so it is invisible to the end user, it appears that I actually have to build a formal "umbrella framework". These are a mystery and discouraged by Apple, but I think I got most of the details worked out with one fragile symlink that I don't understand why it is so fragile. Since we can depend on having 10.4, I can use @loader_path to help. I discovered @rpath requires 10.5. Code drop on this soon.

--HG--
extra : convert_revision : svn%3Ac70aab31-4412-0410-b14c-859654838e24/trunk%405127